test(stochtrace): Rename helpers file to conftest
